requirements.txt
# Django
Django
django-jinja
django-debug-toolbar
django-subdomains

# Python
python-social-auth

# 3-rd party
Jinja2
easy_thumbnails
humanize

# neo4j
py2neo==2.*
neomodel
